Grid: organize package by layout model
Move DashboardGrid into dashboard-grid/, ResizeHandle into shared/.
Split types.ts into per-folder files. Each layout model owns its
component, types, stories, and tests.
