+++ packages/grid/src/dashboard-grid/types.ts
@@ -1,22 +1,7 @@
 /**
- * External dependencies
+ * Internal dependencies
  */
-import type { useDraggable } from '@dnd-kit/core';
-
-// `useDraggable`'s `listeners` and `attributes` types are not exported
-// from `@dnd-kit/core`'s public surface, so derive them from the hook
-// itself rather than via a deep import.
-type DraggableBindings = ReturnType< typeof useDraggable >;
-
-/**
- * Cursor offset reported by the resize handle, in pixels relative to
- * the gesture start. Width and height are independent so the grid can
- * step columns and rows separately.
- */
-export type ResizeDelta = {
-	width: number;
-	height: number;
-};
+import type { ResizeDelta, ResizeHandleRenderProps } from '../shared/types';
 
 /**
  * Dashboard grid layout item definition.
@@ -51,87 +36,6 @@ export type DashboardGridLayoutItem = {
 	 */
 	order?: number;
 };
-
-/**
- * Props received by a custom resize handle component. Spread `listeners`
- * and `attributes` onto the element that should respond to the gesture,
- * and assign `ref` to the same element so dnd-kit can track it.
- */
-export interface ResizeHandleRenderProps {
-	/**
-	 * Ref callback to attach to the gesture-bearing element.
-	 */
-	ref: DraggableBindings[ 'setNodeRef' ];
-
-	/**
-	 * Pointer/keyboard event listeners that initiate the drag.
-	 */
-	listeners: DraggableBindings[ 'listeners' ];
-
-	/**
-	 * Accessibility and dnd-kit attributes (role, aria-*, tabIndex…).
-	 */
-	attributes: DraggableBindings[ 'attributes' ];
-
-	/**
-	 * Whether vertical resizing is allowed for this tile. Useful for
-	 * adapting the cursor or visual cue.
-	 */
-	verticalResizable: boolean;
-
-	/**
-	 * True while the user is actively dragging this handle. Use it to
-	 * swap colors, icons, or transforms during the gesture.
-	 */
-	isResizing: boolean;
-
-	/**
-	 * Owning grid item's `key`. Available so consumers can render
-	 * per-tile content if needed.
-	 */
-	itemId?: string;
-}
-
-/**
- * Props for the internal `<ResizeHandle />` wrapper.
- */
-export interface ResizeHandleProps {
-	/**
-	 * Owning grid item's `key`. Forwarded as `data.itemId` on the
-	 * draggable so the parent can correlate the gesture with a tile
-	 * if needed.
-	 */
-	itemId?: string;
-
-	/**
-	 * Whether the handle should track vertical movement. When false,
-	 * the handle still appears but only emits horizontal deltas, and
-	 * the cursor is constrained to the column resize axis.
-	 *
-	 * @default true
-	 */
-	verticalResizable?: boolean;
-
-	/**
-	 * Callback fired while the handle is being dragged. Receives the
-	 * cursor offset from the gesture start in pixels.
-	 */
-	onResize?: ( delta: ResizeDelta ) => void;
-
-	/**
-	 * Callback fired when the gesture ends.
-	 */
-	onResizeEnd?: () => void;
-
-	/**
-	 * Component that overrides the default corner triangle with a
-	 * custom element. Receives gesture wiring (`ref`, `listeners`,
-	 * `attributes`) plus context. The grid keeps ownership of the
-	 * `<DndContext>` and the throttled delta loop; consumers are only
-	 * responsible for the visual.
-	 */
-	renderResizeHandle?: React.ComponentType< ResizeHandleRenderProps >;
-}
 
 /**
  * Props for the internal `<GridItem />` wrapper.
@@ -180,7 +84,7 @@ export type GridItemProps = {
 
 	/**
 	 * Content rendered above the draggable area that stays interactive
-	 * in edit mode — typically action buttons, menus, or links. While
+	 * in edit mode, typically action buttons, menus, or links. While
 	 * any tile in the grid is being dragged or resized, this content
 	 * is set `inert` so hovers on other tiles can't steal the gesture.
 	 */
