+++ packages/grid/src/index.ts
@@ -1,7 +1,7 @@
-export { DashboardGrid } from './grid';
+export { DashboardGrid } from './dashboard-grid';
+
 export type {
 	DashboardGridLayoutItem,
 	DashboardGridProps,
-	ResizeDelta,
-	ResizeHandleRenderProps,
-} from './types';
+} from './dashboard-grid/types';
+export type { ResizeDelta, ResizeHandleRenderProps } from './shared/types';
