+++ packages/grid/src/shared/resize-handle.module.css
