+++ packages/grid/src/shared/resize-handle.tsx
