+++ tools/eslint/suppressions.json
@@ -1483,12 +1483,12 @@
 			"count": 1
 		}
 	},
-	"packages/grid/src/grid.tsx": {
+	"packages/grid/src/dashboard-grid/index.tsx": {
 		"react-hooks/refs": {
 			"count": 13
 		}
 	},
-	"packages/grid/src/resize-handle.tsx": {
+	"packages/grid/src/shared/resize-handle.tsx": {
 		"react-hooks/immutability": {
 			"count": 1
 		}
